Change structure of look and feel section in settings
